Use println instead of spellcheck.mode depracated warning.
