+++ .autocorrectrc.template
@@ -1,5 +1,6 @@
+rules:
+  spellcheck: 2
 spellcheck:
-  mode: 2
   words:
     # Please do not add a general English word (eg. apple, python) here.
     # Users can add their special words to their .autocorrectrc file by their need.
