+++ autocorrect/.autocorrectrc.default
@@ -6,7 +6,6 @@ rules:
   no-space-fullwidth: 1
   spellcheck: 2
 spellcheck:
-  mode: 2
   words:
     # Please do not add a general English word (eg. apple, python) here.
     # Users can add their special words to their .autocorrectrc file by their need.
